moved few functions to platform package
