+++ common/utils.go
@@ -121,35 +121,6 @@ func LogNetworkInterfaces() {
 	}
 }
 
-func CheckIfFileExists(filepath string) (bool, error) {
-	_, err := os.Stat(filepath)
-	if err == nil {
-		return true, nil
-	}
-
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-
-	return true, err
-}
-
-func CreateDirectory(dirPath string) error {
-	var err error
-
-	if dirPath == "" {
-		log.Printf("dirPath is empty, nothing to create.")
-		return nil
-	}
-
-	isExist, _ := CheckIfFileExists(dirPath)
-	if !isExist {
-		err = os.Mkdir(dirPath, os.ModePerm)
-	}
-
-	return err
-}
-
 func IpToInt(ip net.IP) uint32 {
 	if len(ip) == 16 {
 		return binary.BigEndian.Uint32(ip[12:16])
